fixed Wire.h reading problem

+++ lib/AS5600/AS5600.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <Arduino.h>
-#include <Wire.h>
+#include <driver/i2c.h>
 #include <driver/adc.h>
 
 
@@ -16,14 +16,14 @@ class AS5600 {
     
     private:
         uint8_t analog_pin;
-        uint16_t angle;
+        int angle;
 
     public:
         AS5600 (uint8_t analog_pin);
         void setup();
 
-        uint16_t read_angle();
-        uint16_t get_angle();
+        int read_angle();
+        int get_angle();
 
         uint16_t read_angle_output();
